Revert "Api refactor (#81)"
This reverts commit 53969b6e9bc0c58f31a10068ac03dcd460eb981d.
