+++ classes/member_index.py
@@ -1,4 +1,4 @@
-import api.modules.formattable as format
+import modules.formattable as format
 
 class MemberIndex:
     def __init__(self, data, state):
