+++ test/test_member_index.py
@@ -1,5 +1,5 @@
-from api.app import app
-from api.classes.member_index import MemberIndex
+from app import app
+from classes.member_index import MemberIndex
 
 def test_initialize():
   data = {
@@ -33,3 +33,4 @@ def test_initialize():
   assert member.role == 'Representative'
   assert member.party == 'Democrat'
   assert member.state == 'Colorado'
+
